requirements.txt
asgiref==3.4.1
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.7
dj-database-url==0.5.0
Django==3.2.8
django-filter==21.1
django-on-heroku==1.1.2
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.3
psycopg2-binary==2.9.1
pycodestyle==2.8.0
python-dotenv==0.19.1
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
whitenoise==5.3.0
